+++ Lua/LuaBindingsEntities.cpp
@@ -240,10 +240,12 @@ namespace RTE {
 		.property("MaxInventoryMass", &Actor::GetMaxInventoryMass)
 		.property("MovePathSize", &Actor::GetMovePathSize)
 		.property("MovePathEnd", &Actor::GetMovePathEnd)
+		.property("IsWaitingOnNewMovePath", &Actor::IsWaitingOnNewMovePath)
 		.property("AimDistance", &Actor::GetAimDistance, &Actor::SetAimDistance)
 		.property("SightDistance", &Actor::GetSightDistance, &Actor::SetSightDistance)
 		.property("PieMenu", &Actor::GetPieMenu, &LuaAdaptersPropertyOwnershipSafetyFaker::ActorSetPieMenu)
 		.property("AIBaseDigStrength", &Actor::GetAIBaseDigStrength, &Actor::SetAIBaseDigStrength)
+		.property("DigStrength", &Actor::EstimateDigStrength)
 		.property("SceneWaypoints", &LuaAdaptersActor::GetSceneWaypoints, luabind::adopt(luabind::return_value) + luabind::return_stl_iterator)
 		.property("LimbPushForcesAndCollisionsDisabled", &Actor::GetLimbPushForcesAndCollisionsDisabled, &Actor::SetLimbPushForcesAndCollisionsDisabled)
 
@@ -1249,6 +1251,8 @@ namespace RTE {
 		.def("PathFindingUpdated", &Scene::PathFindingUpdated)
 		.def("CalculatePath", &LuaAdaptersScene::CalculatePath1)
 		.def("CalculatePath", &LuaAdaptersScene::CalculatePath2)
+		.def("CalculatePathAsync", &LuaAdaptersScene::CalculatePathAsync1)
+		.def("CalculatePathAsync", &LuaAdaptersScene::CalculatePathAsync2)
 
 		.enum_("PlacedObjectSets")[
 			luabind::value("PLACEONLOAD", Scene::PlacedObjectSets::PLACEONLOAD),