+++ Lua/LuaBindingsSystem.cpp
@@ -233,4 +233,22 @@ namespace RTE {
 		.def("AbsRotateTo", &Vector::AbsRotateTo)
 		.def("SetXY", &Vector::SetXY);
 	}
+
+/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+	LuaBindingRegisterFunctionDefinitionForType(SystemLuaBindings, PathRequest) {
+		using namespace micropather;
+		return luabind::class_<PathRequest>("PathRequest")
+
+		.def_readonly("Path", &PathRequest::path, luabind::return_stl_iterator)
+		.def_readonly("PathLength", &PathRequest::pathLength)
+		.def_readonly("Status", &PathRequest::status)
+		.def_readonly("TotalCost", &PathRequest::totalCost)
+
+		.enum_("Status")[
+			luabind::value("Solved", micropather::MicroPather::SOLVED),
+			luabind::value("NoSolution", micropather::MicroPather::NO_SOLUTION),
+			luabind::value("StartEndSame", micropather::MicroPather::START_END_SAME)
+		];
+	}
 }