+++ System/PathFinder.h
@@ -12,6 +12,19 @@ namespace RTE {
 	class Material;
 
 	/// <summary>
+	/// Information required to make an async pathing request.
+	/// </summary>
+	struct PathRequest {
+		bool complete = false;
+		int status = MicroPather::NO_SOLUTION;
+		std::list<Vector> path;
+		float pathLength;
+		float totalCost = 0.0f;
+	};
+
+	using PathCompleteCallback = std::function<void (std::shared_ptr<volatile PathRequest>)>;
+
+	/// <summary>
 	/// Contains everything related to a PathNode on the path grid used by PathFinder.
 	/// </summary>
 	struct PathNode {
@@ -96,6 +109,7 @@ namespace RTE {
 #pragma region PathFinding
 		/// <summary>
 		/// Calculates and returns the least difficult path between two points on the current scene.
+		/// This is synchronous, and will block the current thread!
 		/// </summary>
 		/// <param name="start">Start positions on the scene to find the path between.</param>
 		/// <param name="end">End positions on the scene to find the path between.</param>
@@ -103,7 +117,18 @@ namespace RTE {
 		/// <param name="totalCostResult">The total minimum difficulty cost calculated between the two points on the scene.</param>
 		/// <param name="digStrength">What material strength the search is capable of digging through.</param>
 		/// <returns>Success or failure, expressed as SOLVED, NO_SOLUTION, or START_END_SAME.</returns>
-		int CalculatePath(Vector start, Vector end, std::list<Vector> &pathResult, float &totalCostResult, float digStrength = 1);
+		int CalculatePath(Vector start, Vector end, std::list<Vector> &pathResult, float &totalCostResult, float digStrength);
+
+		/// <summary>
+		/// Calculates and returns the least difficult path between two points on the current scene.
+		/// This is asynchronous and thus will not block the current thread.
+		/// </summary>
+		/// <param name="start">Start positions on the scene to find the path between.</param>
+		/// <param name="end">End positions on the scene to find the path between.</param>
+		/// <param name="digStrength">What material strength the search is capable of digging through.</param>
+		/// <param name="callback">The callback function to be run when the path calculation is completed.</param>
+		/// <returns>A shared pointer to the volatile PathRequest to be used to track whehter the asynchrnous path calculation has been completed, and check its results.</returns>
+		std::shared_ptr<volatile PathRequest> CalculatePathAsync(Vector start, Vector end, float digStrength, PathCompleteCallback callback = nullptr);
 
 		/// <summary>
 		/// Recalculates all the costs between all the PathNodes by tracing lines in the material layer and summing all the material strengths for each encountered pixel. Also resets the pather itself.
@@ -164,6 +189,7 @@ namespace RTE {
 		int m_GridHeight; //!< The height of the pathing grid, in PathNodes.
 		bool m_WrapsX; //!< Whether the pathing grid wraps on the X axis.
 		bool m_WrapsY; //!< Whether the pathing grid wraps on the Y axis.
+		std::atomic<int> m_CurrentPathingRequests; //!< The number of active async pathing requests.
 
 		/// <summary>
 		/// Gets the pather for this thread. Lazily-initialized for each new thread that needs a pather.